chore: refactor cors config
