+++ src/common/test.utils.ts
@@ -1,13 +1,11 @@
 import {
   ExecutionContext,
   INestApplication,
-  UnauthorizedException,
   ValidationPipe,
   VersioningType,
 } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import { isURL } from 'class-validator';
 import cookieParser from 'cookie-parser';
+import { corsConfig } from './cors.config';
 
 export function appConfig(app: INestApplication) {
   app.enableVersioning({
@@ -18,35 +16,7 @@ export function appConfig(app: INestApplication) {
   app.use(cookieParser());
   app.useGlobalPipes(new ValidationPipe());
 
-  const configService = app.get(ConfigService);
-  const allowedOrigins = configService.get('ALLOWED_ORIGINS').split(',');
-  for (const origin of allowedOrigins) {
-    if (!isURL(origin)) {
-      throw new Error(`Origin ${origin} is not a valid URL`);
-    }
-  }
-  const restrictCorsOrigins = configService.get<boolean>(
-    'RESTRICT_CORS_ORIGINS'
-  );
-  app.enableCors({
-    credentials: true,
-    origin: restrictCorsOrigins
-      ? function (origin, callback) {
-          if (origin) {
-            if (allowedOrigins.includes(origin)) {
-              callback(undefined, origin);
-            } else {
-              callback(
-                new UnauthorizedException(`Origin ${origin} is not allowed`),
-                false
-              );
-            }
-          } else {
-            callback(undefined, true);
-          }
-        }
-      : true,
-  });
+  corsConfig(app);
 }
 
 export const MockJWTAuthGuard = {
