+++ src/main.ts
@@ -4,18 +4,14 @@ loadEnvVars(); // this is required by the Auth decorator depending on process.en
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import {
-  UnauthorizedException,
-  ValidationPipe,
-  VersioningType,
-} from '@nestjs/common';
+import { ValidationPipe, VersioningType } from '@nestjs/common';
 import cookieParser from 'cookie-parser';
 import { ConfigService } from '@nestjs/config';
 import compression from 'compression';
 import helmet from 'helmet';
 import { SentryService } from './modules/sentry/sentry.service';
 import * as SentryNode from '@sentry/node';
-import { isURL } from 'class-validator';
+import { corsConfig } from './common/cors.config';
 
 // This allows TypeScript to detect our global value and properly assign maps for sentry
 // See more here: https://docs.sentry.io/platforms/node/typescript/
@@ -75,34 +71,7 @@ async function bootstrap() {
   const document = SwaggerModule.createDocument(app, builtOptions);
   SwaggerModule.setup('api', app, document);
 
-  const allowedOrigins = configService.get('ALLOWED_ORIGINS').split(',');
-  for (const origin of allowedOrigins) {
-    if (!isURL(origin)) {
-      throw new Error(`Origin ${origin} is not a valid URL`);
-    }
-  }
-  const restrictCorsOrigins = configService.get<boolean>(
-    'RESTRICT_CORS_ORIGINS'
-  );
-  app.enableCors({
-    credentials: true,
-    origin: restrictCorsOrigins
-      ? function (origin, callback) {
-          if (origin) {
-            if (allowedOrigins.includes(origin)) {
-              callback(undefined, origin);
-            } else {
-              callback(
-                new UnauthorizedException(`Origin ${origin} is not allowed`),
-                false
-              );
-            }
-          } else {
-            callback(undefined, true);
-          }
-        }
-      : true,
-  });
+  corsConfig(app);
 
   app.use(SentryNode.Handlers.errorHandler());
   await app.listen(configService.get('NESTJS_PORT'));
